Suppression du modèle GP_Pilotes et création d'un nouveau modèle avec des relations mises à jour
